feat(budget): feat: support SSR prefetch with react-query by reading auth token from cookies() in server component
